Moves command's functions to a designated module
