+++ scrapyd_client/commands.py
@@ -1,0 +1,45 @@
+import sys
+from textwrap import indent
+
+from scrapyd_client import lib
+
+
+INDENT_PREFIX = '  '
+
+
+def deploy(args):
+    """ Deploys a Scrapy project to a Scrapyd instance.
+        For help on this command, invoke `scrapyd-deploy`. """
+    from scrapyd_client import deploy
+    sys.argv.pop(1)
+    deploy.main()
+
+
+def projects(args):
+    """ Lists all projects deployed on a Scrapyd instance. """
+    _projects = lib.get_projects(args.target)
+    if _projects:
+        print('\n'.join(_projects))
+
+
+def schedule(args):
+    """ Schedules the specified spider(s). """
+    job_args = dict((x[0], x[1]) for x in (y.split('=', 1) for y in args.arg))
+    _projects = lib.get_projects(args.target, args.project)
+    for project in _projects:
+        _spiders = lib.get_spiders(args.target, _projects, args.spider)
+        for spider in _spiders:
+            job_id = lib.schedule(args.target, project, spider, job_args)
+            print(job_id)
+
+
+def spiders(args):
+    """ Lists all spiders for the given project(s). """
+    _projects = lib.get_projects(args.target, args.project)
+    for project in _projects:
+        print('{}:'.format(project))
+        _spiders = lib.get_spiders(args.target, project)
+        if _spiders:
+            print(indent('\n'.join(_spiders), INDENT_PREFIX))
+        else:
+            print(INDENT_PREFIX + 'No spiders.')
